requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
requests==2.31.0
